backedned/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless>=4.9.0.80
librosa>=0.10.1
numpy>=1.24.0
Pillow>=10.0.0
soundfile>=0.12.1
huggingface-hub>=0.23.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
